feat: T2 prerequisites — clean Streamlit leaks from domain code
T2-01: gitignore secrets.toml + pre-commit hook
T2-02: remove import streamlit from integrations/config, add cache_decorator.py
T2-03: brand filter preserves Grupo row
T2-07: vectorize shortages.py (np.where for TimeDelta)
T2-08: replace .apply(lambda) in donotbuy.py with list comprehension
T2-09: add codigos_visible lazy filter parameter
T2-10: date formats verified as dd-MM-yyyy
T2-12: standardize secret key names
T2-13: add CSS class constants to constants.py
T2-16: rename master_products → df_master_products

+++ orders_master/integrations/shortages.py
@@ -50,7 +50,7 @@ def fetch_shortages_db(url: str, codigos_visible: set[int] | None = None) -> pd.
         delta.notna(),
         delta.dt.days,
         pd.NA,
-    )
+    ).astype("Int64")
 
     if codigos_visible is not None:
         df = df[
